// src/apis/generated/model/notificationRes.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * 동행 API
 * 동행 서비스 API 입니다.
 * OpenAPI spec version: v1
 */

export interface NotificationRes {
  notificationId?: number;
  message?: string;
  isRead?: boolean;
  createdAt?: string;
  userProcedureChecklistId?: number;
  deceasedProfileId?: number;
  deceasedName?: string;
  dueDate?: string;
}